options: Handle command line option errors more gracefully
Stop printing the "BUG: unprocessed option id -2" message when user
passes wrong arguments.

Before this change:

    $ ./build/release/ponyc -V
    ./build/release/ponyc: '-V' option requires an argument!
    BUG: unprocessed option id -2
    $ echo $?
    255

After this change:

    $ ./build/release/ponyc -V
    ./build/release/ponyc: '-V' option requires an argument!
    $ echo $?
    255

+++ src/libponyrt/options/options.c
@@ -247,5 +247,5 @@ int ponyint_opt_next(opt_state_t* s)
 
   strip_accepted_opts(s);
 
-  return m->id;
+  return (int)m->id;
 }
